[LM-84] Split test_server.py into per-area test modules under tests/
