+++ tests/test_board.py
@@ -1,0 +1,47 @@
+import server
+
+
+def test_board_returns_list(shared_client):
+    server._insert_verdict(server.VerdictPayload(
+        project="proj-b", role="reviewer", model="gpt-4", points=5, reason="ok"
+    ))
+    resp = shared_client.get("/api/board")
+    assert resp.status_code == 200
+    data = resp.json()
+    assert isinstance(data, list)
+    entry = next((r for r in data if r["project"] == "proj-b"), None)
+    assert entry is not None
+    assert entry["total_points"] == 5
+
+
+def test_board_cumulative_scores(shared_client):
+    server._insert_verdict(server.VerdictPayload(
+        project="proj-d", role="planner", model="claude-3", points=8, reason="first"
+    ))
+    server._insert_verdict(server.VerdictPayload(
+        project="proj-d", role="planner", model="claude-3", points=7, reason="second"
+    ))
+    resp = shared_client.get("/api/board")
+    data = resp.json()
+    entry = next(r for r in data if r["project"] == "proj-d" and r["role"] == "planner")
+    assert entry["total_points"] == 15
+    assert entry["verdict_count"] == 2
+
+
+def test_api_verdicts_empty(isolated_client):
+    response = isolated_client.get("/api/verdicts")
+    assert response.status_code == 200
+    assert response.json() == []
+
+
+def test_api_verdicts_after_post(isolated_client):
+    isolated_client.post("/api/verdict", json={
+        "project": "test", "role": "builder", "points": 10, "reason": "good work"
+    })
+    response = isolated_client.get("/api/verdicts")
+    assert response.status_code == 200
+    data = response.json()
+    assert len(data) == 1
+    assert data[0]["role"] == "builder"
+    assert data[0]["points"] == 10
+    assert data[0]["reason"] == "good work"
